bring all reference dep as test dependencies
