+++ project/plugins.sbt
@@ -3,3 +3,5 @@ addSbtPlugin("org.foundweekends" % "sbt-bintray" % "0.5.5")
 addSbtPlugin("com.github.gseitz"        % "sbt-release"             % "1.0.11")
 
 addSbtPlugin("io.get-coursier"     %  "sbt-coursier"           % "1.0.3")
+
+addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.4.1")
